+++ src/api/cli/send.go
@@ -56,7 +56,7 @@ func sendCmd() gcli.Command {
 		Action: func(c *gcli.Context) error {
 			rpcClient := RpcClientFromContext(c)
 
-			rawtx, err := createRawTx(c)
+			rawtx, err := createRawTxCmdHandler(c)
 			if err != nil {
 				errorWithHelp(c, err)
 				return nil
@@ -101,5 +101,4 @@ func SendFromAddress(c *webrpc.Client, addr, walletFile, chgAddr string, toAddrs
 	}
 
 	return c.InjectTransaction(rawTx)
-
 }