+++ src/visor/blockdb/block_tree_test.go
@@ -4,10 +4,11 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	"github.com/spaco/spo/src/cipher"
 	"github.com/spaco/spo/src/coin"
 	"github.com/spaco/spo/src/testutil"
-	"github.com/stretchr/testify/assert"
 )
 
 type blockInfo struct {