+++ src/visor/blockdb/blocksigs.go
@@ -2,6 +2,7 @@ package blockdb
 
 import (
 	"github.com/boltdb/bolt"
+
 	"github.com/spaco/spo/src/cipher"
 	"github.com/spaco/spo/src/cipher/encoder"
 	"github.com/spaco/spo/src/visor/bucket"