+++ src/visor/historydb/transaction.go
@@ -7,6 +7,7 @@ package historydb
 
 import (
 	"github.com/boltdb/bolt"
+
 	"github.com/spaco/spo/src/cipher"
 	"github.com/spaco/spo/src/cipher/encoder"
 	"github.com/spaco/spo/src/coin"