+++ src/visor/readable_test.go
@@ -2,11 +2,11 @@
 package visor
 
 import (
-	"crypto/rand"
+	"testing"
 	"time"
 
-	"github.com/spaco/spo/src/cipher"
 	"github.com/spaco/spo/src/coin"
+	"github.com/spaco/spo/src/testutil"
 	"github.com/spaco/spo/src/util/utc"
 )
 
@@ -106,16 +106,10 @@ const (
 // 	}
 // }
 
-func randSHA256() cipher.SHA256 {
-	b := make([]byte, 128)
-	rand.Read(b)
-	return cipher.SumSHA256(b)
-}
-
-func createUnconfirmedTxn() UnconfirmedTxn {
+func createUnconfirmedTxn(t *testing.T) UnconfirmedTxn {
 	ut := UnconfirmedTxn{}
 	ut.Txn = coin.Transaction{}
-	ut.Txn.InnerHash = randSHA256()
+	ut.Txn.InnerHash = testutil.RandSHA256(t)
 	ut.Received = utc.Now().UnixNano()
 	ut.Checked = ut.Received
 	ut.Announced = time.Time{}.UnixNano()